# specs/001-backend-task-service/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Backend Task Service API
  version: 1.0.0
  description: |
    RESTful API for Todo task management with user-scoped data access.
    All endpoints enforce user isolation via user_id parameter.
  contact:
    name: Todo AI Chatbot Project

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: tasks
    description: Task CRUD operations

paths:
  /api/{user_id}/tasks:
    get:
      tags: [tasks]
      summary: List all tasks for a user
      operationId: listTasks
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user who owns the tasks
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskRead'
      security: []

    post:
      tags: [tasks]
      summary: Create a new task
      operationId: createTask
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user who will own the task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security: []

  /api/{user_id}/tasks/{task_id}:
    get:
      tags: [tasks]
      summary: Get a single task
      operationId: getTask
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user requesting the task
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task to retrieve
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '404':
          description: Task not found or belongs to different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []

    put:
      tags: [tasks]
      summary: Update a task
      operationId: updateTask
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user who owns the task
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '404':
          description: Task not found or belongs to different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security: []

    delete:
      tags: [tasks]
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user who owns the task
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task to delete
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '404':
          description: Task not found or belongs to different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []

  /api/{user_id}/tasks/{task_id}/complete:
    patch:
      tags: [tasks]
      summary: Toggle task completion status
      operationId: toggleTaskComplete
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: The ID of the user who owns the task
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the task to toggle
      responses:
        '200':
          description: Task completion toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '404':
          description: Task not found or belongs to different user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security: []

components:
  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title
        description:
          type: string
          maxLength: 5000
          description: Optional task description

    TaskRead:
      type: object
      required:
        - id
        - user_id
        - title
        - is_completed
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique task identifier
        user_id:
          type: string
          description: ID of the user who owns this task
        title:
          type: string
          description: Task title
        description:
          type: string
          nullable: true
          description: Optional task description
        is_completed:
          type: boolean
          description: Task completion status
        created_at:
          type: string
          format: date-time
          description: ISO 8601 creation timestamp
        updated_at:
          type: string
          format: date-time
          description: ISO 8601 last modification timestamp

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: New task title
        description:
          type: string
          maxLength: 5000
          description: New task description

    DeleteResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Task deleted successfully"

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          example: "Task not found"

    HTTPValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            required:
              - loc
              - msg
              - type
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string